Extract EventPage and EventOverviewPage components
