+++ pages/events/[slug].tsx
@@ -1,17 +1,8 @@
 import type { NextPage, GetStaticPaths, GetStaticProps } from "next";
 import { PortalEvent, PortalProject, PortalUser } from "lib/portal-types";
-import * as Typography from "components/typography";
-import { Layout, Section, SectionContent } from "components/layout";
-import * as S from "components/event/styles";
-import EventCard from "components/event/event-card";
-import { Link } from "components/links";
-import { CardRow } from "components/layout";
-import { getResizedImgUrl } from "lib/utils";
-import RenderMarkdown from "components/markdown";
 import { ParsedUrlQuery } from "querystring";
 import { siteData } from "lib/site-data";
-import { Route } from "lib/routing";
-import strings from "content/strings.json";
+import EventPage from "components/event/page";
 
 interface PageProps {
   event: PortalEvent;
@@ -26,75 +17,7 @@ interface QueryParams extends ParsedUrlQuery {
 }
 
 const Page: NextPage<PageProps> = (props) => {
-  const { event, project, projects, owner, events } = props;
-  const coverImageUrl = event.coverImageUrl || project.coverImageUrl;
-  const getEventProject = (e: PortalEvent) =>
-    projects.find((p) => p.id === e.projectId)!;
-  const otherEvents = events.filter(
-    (e) => e.status === "live" && e.id !== event.id
-  );
-  return (
-    <Layout
-      crumbs={[
-        { path: Route.dashboard, label: strings.crumbs.dashboard },
-        { label: event.name },
-      ]}
-      head={{
-        title: event.name,
-        description: event.summary,
-        coverUrl: coverImageUrl,
-      }}
-    >
-      <Section>
-        <SectionContent>
-          <Typography.Heading1>{event.name}</Typography.Heading1>
-          <Typography.Body>{event.summary}</Typography.Body>
-          <S.CoverImageWrapper>
-            <S.CoverFilter />
-            <S.CoverImage
-              src={getResizedImgUrl(coverImageUrl, 1160)}
-              loading="lazy"
-            />
-          </S.CoverImageWrapper>
-        </SectionContent>
-      </Section>
-      <Section>
-        <SectionContent>
-          <S.TwoColumnLayout>
-            <S.MainColumn>
-              <Typography.Body>
-                <RenderMarkdown source={event.description} />
-              </Typography.Body>
-            </S.MainColumn>
-            <S.ReminderColumn>
-              <EventCard event={event} owner={owner} project={project} />
-            </S.ReminderColumn>
-          </S.TwoColumnLayout>
-        </SectionContent>
-      </Section>
-      <Section>
-        <SectionContent>
-          <S.Container>
-            <S.CategoryHeader>
-              <S.Title>Další akce</S.Title>
-              <Link to={Route.dashboard}>Všechny příležitosti</Link>
-            </S.CategoryHeader>
-            <S.CardWrapper>
-              <CardRow>
-                {otherEvents.slice(0, 3).map((event, index) => (
-                  <S.ProjectCard
-                    key={index}
-                    event={event}
-                    project={getEventProject(event)}
-                  />
-                ))}
-              </CardRow>
-            </S.CardWrapper>
-          </S.Container>
-        </SectionContent>
-      </Section>
-    </Layout>
-  );
+  return <EventPage {...props} />;
 };
 
 export const getStaticPaths: GetStaticPaths<QueryParams> = async () => {
